requirements-dev.txt
# Development dependencies for EconoNet
# Install with: pip install -r requirements-dev.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
responses>=0.23.0

# Code formatting and linting
black>=23.0.0
ruff>=0.0.280
mypy>=1.5.0
pre-commit>=3.3.0

# Notebook support
papermill>=2.4.0
nbconvert>=7.7.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.1.0
mkdocstrings[python]>=0.22.0

# Type stubs
types-requests>=2.31.0
pandas-stubs>=2.0.0

# Development tools
ipython>=8.12.0
jupyter-lab>=4.0.0
